+++ src/main/java/com/example/service/ICategoriaService.java
@@ -7,5 +7,5 @@ import com.example.entities.Categoria;
 public interface ICategoriaService {
 public List<Categoria>getCategorias();    
 public void guardar(Categoria categoria);
-public Categoria getCategoria (long idCategoria);     
+public Categoria getCategoria (int idCategoria);     
 }
