+++ src/main/java/com/example/service/IViajeService.java
@@ -5,11 +5,10 @@ import java.util.List;
 import com.example.entities.Viaje;
 
 public interface IViajeService {
+    
     public List<Viaje> getViajes();
-
-    public Viaje getViaje(long idViaje);
-
+    public Viaje getViaje(int idViaje);
     public void guardar(Viaje viaje);
-
-    public void eliminar(long idViaje);
+    public void eliminar(int idViaje);
 }
+
